Decouple sequence-field's format from its in-memory types (#18519)
## Description

Decouples the types used to represent sequence-field changes in memory
from those we send over the wire. This change was relatively mechanical
but is likely easier to review from a 'fresh code' standpoint, with the
exception of moving the `DiscriminatedUnionDispatcher` into the `codec`
library as it's a generally useful encoding for optional unions.

Notable features:
- Moves typebox schema types for sequence-field to a separate file from
the in-memory types. This noticeably cleans up the dependency structure
of sequence-field (only the codec file depends on the encoded format)
- More strongly-types the codec, avoiding using spread operators when
encoding data.
- Leverages the discriminated union encoding outlined in
`discriminatedUnions.ts` for marks

There are still some planned changes to make sequence-field's encoding
more efficient, so this doesn't attempt to settle all of the encoding
details. However, it should make future changes much more local, and
also brings the organization of the code more inline with optional/value
field.

## Breaking Changes

This is a breaking change for the persisted sequence-field format. No
back-compat effort has been made.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>