build(example-utils): Add release tag enforcement (#18571)
Follow-up to #18411, which added explicit `@internal` release tags to
all API members. This PR adds "linter" enforcement that release tags are
compatible between this package and its local (fluid) dependencies.

Note: example-utils was the only `examples` package already configured
with `api-extractor`. If we want to add this enforcement to the other
packages here (which we likely do, though with low urgency due to all of
these being repo-private), we'll need to do a pass to add api-extractor
too.

Also removes an unused `api-extractor.json` from
`examples/data-objects/todo`.