tree2: Remove support for simple-tree with flex-tree-schema outside of tree2 package (#18587)
## Description

All out of package use of tree2 now has to pick between the:

- simple-tree abstraction using the new class based schema. This focuses
on node kinds, has List nodes etc. and exposes minimal functionality and
leverages implicit defaults. This is marked `@beta`
- flex-tree abstraction using schema-builder-base. This focuses on nodes
+ fields and FieldKinds, and exposes maximal functionality and is more
explicit. This is marked `@alpha`.

This change also fixes several doc links which used to go to the wrong
places (ex: from the beta API to the alpha API) but were not caught
until those members were removed from the API.

This change specifically avoids renames and non-required doc updates.
Ex: NodeBase should probably take over the no longer exported TreeNode
type name, but this just updates links to it, and does not do any of the
renames. This should help keep this PR focused and easier to merge and
review.

## Breaking Changes

Users of simple-tree with flex-tree schema will have to port to the new
SchemaFactory (from SchemaBuilder).

Users of flex-tree with have to port from SchemaBuilder to
SchemaBuilderBase, and directly refer to TreeFieldSchema.create when
needed.