Add factories for SharedTree lists and maps (#18595)
## Description

This adds the ability to construct lists and maps via `new ...` in the
class-tree layer. This does not add `.create` for lists and maps in the
simple-tree layer.

## Reviewer Guidance

* Raw nodes are implemented as FlexTreeNodes (including raw object
nodes, which have been re-implemented). Much of the implementation is
stubbed out since it is not needed, but if/when we decide to allow
reading and writing to raw proxies, it can be implemented by filling out
those classes/interfaces.
* Proxy content extraction has changed structure a little bit. Now that
extraction and hydration is occurring for all three types of nodes
rather than just objects, some common behavior has been factored out
into helper methods.