build(routerlicious): Update existing api-extractor configs to use newest base config and associated patterns (#18574)
Also enabled cross-package release tag validation, as a follow up to
#18413.

Notes:
* This PR only updates the packages that are already using API-Extractor
as a part of their build. We'll want to do a separate pass to add this
infra to the packages that are not yet using it.
* This PR adds a copy of the api-extractor pnpm patch used by the client
release group and others (which lives at the repo root) to the
routerlicious directory. This was added due to routerlicious' CI build
being run in docker and needing to copy around files. If the CI infra
here changes in the future, this should be deduplicated.