Make OptionalField child changes be keyed on input register id (#18600)
## Description

Converts OptionalField's format to refer to child node changes using the
RegisterId that the node occupies in the input context of the change
rather than the output context.
This makes it consistent with the delta format (which is not necessary,
but convenient) and generally simplifies some of the ChangeRebaser
logic.

## Breaking Changes

This is a breaking change to OptionalField's format semantics, since
child changes refer to the node occupied by a register in the input
context rather than the output context. I was mildly disturbed that this
didn't already result in snapshot test failure, so I added a snapshot
scenario which exercises most of optional-field's interesting scenarios
from the edit manager perspective (this snapshot test would have been
broken had it existed before this change). As usual, no back-compat
attempt was made as the format is not yet stable.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>