feat(eslint-config-fluid): Add eslint rule to prevent imports of @internal items (#18296)
[AB#5695](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5695)


## Description

Fluid Framework needs to maintain a clear boundary between public and
internal APIs. The usage of @internal and similar annotations is crucial
for this, but we lacked automated enforcement. This PR adds a safeguard
through an eslint rule that ensures only the intended, stable, and
documented parts of the library's API are used in public-facing code
samples and tests. There aren't any custom rules so new infra was added.

When implemented, this eslint rule would rely on including the necessary
tags that we want to prevent the import of, ie: @internal, @alpha etc.
The exceptions property can be used to allow for internal imports from a
specific package.