fix: Website build (#18628)
Our blanket tagging of all API-members as (temporarily) `@internal` had
the unfortunate side-effect of (again, temporarily) breaking our API
docs generation for the website. In particular, we had logic that was
special casing certain API members of certain packages to bundle items
directly other packages' API docs.

We now have the option of doing this in a more general way via
functionality of api-extractor. We will light that functionality up for
`fluid-framework` and `@fluidframework/azure-client` in future PRs. For
now, this PR removes the special casing to fix the website build.