Implement First Write Win for Shared Matrix (#18018)
## Description

ADO Item: [Complete implementation of First writer win ability for
shared matrix
dds.](https://dev.azure.com/fluidframework/internal/_workitems/edit/5958)

Doc link for code understanding:
https://microsoft-my.sharepoint-df.com/:w:/p/jatgarg/ET_qkVeiO95DsioiFXZe7mwB17LeCQE-rVvnC2y8aCTM9w?e=39OOvv

Implement First Write Win Policy for Shared Matrix DDS. 

**Scenario 1: Win case:** 
Client tries to set value in an empty cell with no race condition where
no other client is trying to set the value at the same time, then this
op will win, and all clients will set this value in the cell.

**Scenario 2: Race Condition:** 
More than 1 client tries to set the value in a cell and has not observed
any set op from any other client, the client will send the ops with
local changes in the shared matrix. Op which is sequenced first will
win. For a client with local changes, but the op from another client
wins, we will overwrite the local changes and ignore the op from itself.
We will raise an event and give new and old value to the losing client,
and it can resolve the object/contents of the cell.

**Scenario 3: Overwrite:** 
In this scenario, the client tries to set a cell with the knowledge of
the previously acked contents. In this case, the reference sequence
number of the current set is greater than equal to the sequence number
of the last set op. In this case, we overwrite, and the new op will win.
We differentiate between Race Condition and Overwrite case. We track the
seq number of the last set OP for a particular cell which wins. This
helps us to determine whether we want to apply for the new set op or not
based on comparing the ref seq number with seq. If ref seq >= seq, then
update, otherwise it is Race condition scenario 2.

**Scenario 4: Consecutive write before first write is acked:** 
In this case, we raise conflict locally on consecutive write at client
side. The client on this conflict can modify the existing
contents/object instead of resetting it.

**Scenario 5: Offline mode:** 
In this scenario, let us say client creates a set op but gets offline
before sending that op. When it comes online, it catches up and receives
a set op for that cell and applies that due to FWW. In this case, we do
not need to send the op on reconnect but rather raise an event to the
client and let it resolve the conflict just like Race condition Scenario
2.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>