build(client): Add are-the-types-wrong check to client packages (#18650)
This PR adds [the
CLI](https://github.com/arethetypeswrong/arethetypeswrong.github.io/tree/main/packages/cli)
from <https://arethetypeswrong.github.io> to our client packages and
adds a script that runs it against the package. The check is slow
because it packs the package first and runs on the packed contents.

Because the check is slow, it does not run automatically during local
builds but can be triggered manually. In CI the checks run as a test
since they rely on the build output. Ideally this check would be done in
the "pack" phase since we are already packing packages there and could
run on the packed content directly. But that is a bigger CI-centric
change, so I opted not to do that right now.

There are several azure projects that are [producing invalid ESM
types](https://github.com/arethetypeswrong/arethetypeswrong.github.io/blob/main/docs/problems/FalseCJS.md).
Those failures are ignored for now using .attw.json config files in the
individual projects - they'll be cleaned up as we fix the types/exports
field for those packages.

This PR also removes the exports field from most packages, so that we
can re-enable exports in a controlled fashion, while being protected
from releasing bad types by attw.