feat(api-markdown-documenter): Allow filtering generated docs based on minimal release level (#18663)
This PR adds the `minimalReleaseLevel` configuration option, which
allows consumers to specify a release tag to serve as the "minimal"
level of API items to be processed when generating the API docs suite.

E.g. A user could specify `public` to only generate docs for the public
API surface, or they could specify `beta` to include the API surface
containing beta and public members (but not alpha or internal).

The motivation here is to enable hiding `@alpha` API members from the
public (fluidframework.com) documentation, as those APIs are (currently)
designated for pre-approved Microsoft-internal use only. Our internal
API docs site can continue to show alpha docs.

This PR also updates a couple of the end-to-end test configurations to
leverage this new functionality - you can see the resulting changes in
the test snapshots.

AB#6469