Log sanitization during service start config dump (#18670)
## Description

This PR adds a feature to log the entire service config during start of
the service. During the logging, we want to redact any sensitive
information from the logged config while not altering the config object
itself.

To redact secrets, we use `fast-redact`. We add a new optional parameter
`secretsList`. The `ConfigDumper` object has a default list. If no list
is provided, the default list will be used for redaction in case the
values are found in the config. In case the user provides a list, it
will be appended to the default list. The logging is done only if
`config.configDumpEnabled` is set to true.

## Reviewer Guidance

1. Whether to use the `fast-redact`
(https://www.npmjs.com/package/fast-redact) library or write our own
recursive function to redact secrets before logging.
2. Creating a deep copy of the config object before logging to prevent
the original config from getting corrupted.

### Unit test coverage

1) Ensure that the values in `secretNamesToRedactInConfigDump` are
correctly redacted.
2) Ensure that the redacted object is not the same as the original
config object if a value is redacted.
3) Ensure that the redacted object is not the same as the original
config object if no value is redacted.
4) Should not throw an error if `secretNamesToRedactInConfigDump` values
are not present in `nconf`
5) Should not throw an error if `secretNamesToRedactInConfigDump`
contains duplicate values
6) Should redact keys inside nested objects in
`secretNamesToRedactInConfigDump`

### FRS changes

I plan to add the `config` object under the frs-shared configmap. This
way it will be easy to maintain access every service. Additionally, I
will update the shell extension files to have a list of default secrets
to redact. This can be changed as per our needs.

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).