Decouple Fluid-Static and the Service Clients From IFluidDataStoreFactory  (#18688)
This change decouples DataObjectClass from the IFluidDataStoreFactory
but still requires a type of the same shape. This will help reduce the
amount of our surface area we need to make public when we make
AzureClient public. This is a non-breaking change for existing usages,
and still allows the usage of DataObject which are an undocumented
feature.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>