More compile-time variable fixes.
The 'buildBase' variable is kind of redundant; it's always true for all server pipelines except gitssh (which also doesn't provide any of the related test/lint/pack/docs parameters), so it was easier to get rid of it entirely and special case gitssh. 'testCoverage' was only used for one purpose so inlining it was the easiest solution.