Patch tsc-multi to support rewriting imports in declarations (#18712)
This PR patches the repo version of tsc-multi with the contents of this
PR: https://github.com/tylerbutler/tsc-multi/pull/1

This change will enable tsc-multi to rewrite the local imports in
declaration (.d.ts) files just like it does for js files. This is opt-in
behavior in the tsc-multi config.

tsc-multi will not rename the declaration files so that will need to be
done separately (see the changes in
https://github.com/microsoft/FluidFramework/pull/18645 for examples of
how we're going to do the renames).