Added GC op which will be sent with a list of sweep ready nodes when sweep runs (#18632)
## Description
This PR adds a new container runtime op type: GC. This is specific to
garbage collection and will be sent when GC runs. It has a sub-type to
categorize it within the garbage collector. There is currently only one
sub-type `Sweep` that contains the list of nodes that are ready to be
deleted. The format is made flexible to support more GC op types in the
future.

This PR updates the sweep phase flow to the following during
summarization:
1. When sweep phase runs, it identifies nodes that are sweep ready,
i.e., ready to be deleted.
2. The sweep ready data store and blob ids are sent in a GC op (sub-type
Sweep).
3. When the op is received, the sweep ready nodes will be deleted.
3.1. Before this PR, the sweep ready nodes are deleted in the sweep
phase itself.
3.2. Note that all clients that receive this op will delete the sweep
ready nodes.
4. The deleted node list will be updated.
5. The next summary / GC tree will not include the deleted nodes.
5.1. Before this PR, the deleted nodes would be removed frome the GC
tree in the same summary where sweep runs and removed from the summary
tree in the next summary.

With this new flow, all clients will be up-to-date with the state of
deleted nodes and can ensure that deleted objects are not used and do
not cause document corruption - Currently, its possible that a client
has the deleted data store loaded (due to a bug) and keeps sending ops
for it until it loads with a summary that has the object deleted.