tree: Clarify and improve schema mutability contract (#18739)
## Description

This PR updates the in-memory schema repository's mutability contract.
It ensures that data is consistently and cleanly copied. Additionally,
it introduces efficient cloning in some cases. This improvement will
enhance branching efficiency, especially in scenarios where frequent
forking of the schema repository occurs. It also aims to reduce bugs
related to misunderstandings about the mutability of in-memory schema
repository objects.