GC: Ignore incoming changes to data store that has been deleted (#18731)
After a data store is deleted as a result of sweep, further modification
to it must be prevented to cause any data loss. This has been added via
a GC sweep op which will be received by all clients
https://github.com/microsoft/FluidFramework/pull/18632.
However, due to an unexpected bug, a client could have a (to be) deleted
data store loaded and before it receives the GC sweep op, this data
store can send an op or signal. These ops must be ignored after the data
store is deleted to not cause document corruption. This PR adds this
logic. An error is still logged and it can be used to identify if such
case ever happens.

AB#6113