Refactor of IntervalCollection.change and IntervalCollection.changeProperties  (#18676)
Refactors the change method to include the functionality of changing the
endpoints as well as the properties of an interval. It also updates the
signature of the (new) change method to align with the changes made to
the add signature in #17165. The changeProperties method and the old
signature of change will be deprecated and eventually removed.

AB#5738

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>