GC: Simplify configuration for when to enable sweep and for which documents (#18725)
Fixes AB#6481.

In light of #18559, Tombstone and Sweep should always be enabled in
lockstep in terms of which documents are eligible. Today this is
determined by two independent options `gcTombstoneGeneration` and
`gcSweepGeneration`. We are consolidating on a single `gcGeneration`
(with a back-compat measure for old docs with only
`gcTombstoneGeneration`).

The semantics follow the same concept: Any document stamped with the
matching generation value that's current passed will be eligible for
both the Tombstone and Delete stages of Sweep Phase. Importantly, if no
value is passed in, Sweep is considered enabled for that document.

Then in order to configure whether to run the Delete stage in a given
session, there's a new GC option called `enableGCSweep` which can be set
to `true` (default is don't delete, only Tombstone).

### Additional behavior change:

* Small bug fix: Tombstone Enforcement was always false in the first
session for a container if a tombstoneGeneration option was passed