feat(generate:buildVersion): Add support for RC versions (#18373)
Adds support for the 2.0 release candidate version scheme. To make the
implementation simpler, the RC version scheme is the same as the
internal version scheme but uses "rc" as the prerelease identifier
instead of "internal". For example, `2.0.0-rc.1.0.0` will be the first
RC version.

Pre-release/dev builds of RC versions will be of the form
`2.0.0-dev-rc.1.0.0.123456`.

AB#6142