Typing Cleanups Before Public Tagging (#18803)
## Remove Unused IFluidDataStoreRegistry From IFluidDataStoreRuntime

IFluidDataStoreRuntime optionally extended IFluidDataStoreRegistry. This
is never used, so is removed. As with all provider interfaces, consumers
can continue to extend the interface if they have a use and use
FluidObject to inspect for it existing.


## SharedObject processGCDataCore now takes IFluidSerializer rather than
SummarySerializer

This change should be a no-op for consumers, and SummarySerializer and
IFluidSerializer expose the same consumer facing APIs. This change just
makes our APIs more consistent by only using interfaces, rather than a
mix of interfaces and concrete implementations.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>