build(api-markdown-documenter): Depend on `eslint-config-fluid` via file dependency (#18799)
Allows the package to always use the latest and greatest, without
needing to publish a new version of the lint config package and install
it. Also makes testing changes in the config package easier. If we like
this pattern, this will be the first such PR of potentially many :)

Also updates code to address new linter rules