Api 2.0 build docs pipeline (#18677)
Updated build-docs pipeline to generate API json artifacts and markdown
for both 2.0 and 1.0 content.

Currently, the 1.0 api content is generated locally by running
api-extractor on the lts branch content and uploaded to the
fluidframework container storage as 'latest-1.0.tar.gz'. The pipeline
has been updated to download this blob before running the markdown
documenter.

In docs/package.json, a new build:multi-version script was added which
calls the relevant build-docs scripts such as rollup-api-json and
render-api-documentation with an additional parameter boolean parameter
to signal the scripts to process content for previous version apis as
well. (note that, the build-docs pipeline is currently set to run the
old build script until 2.0 release)

Note that the previous branch versions are maintained in two separate
areas since there is no straighforward way to import pipeline parameters
from another yaml file. The first is docs/data/versions.yaml and the
second is the parameters in build-docs.yaml. Some checks will need to be
added to ensure that the two lists match up.