Initialize GC state as soon as possible and update it on (re)connections (#18758)
Currently, the GC state from the base snapshot is initialized when the
container transitions to write mode. This means that the unreferenced
tracking of nodes do not begin until container is in write mode. There
are a couple of implications of this:
1. Any loads / changes to inactive, tombstone or sweep ready objects
before container transitions to write mode are not detected.
2. If a container receives a GC sweep op before it transitions to write
mode, it will delete sweep ready nodes. However, when GC is initialized,
it will start tracking these deleted nodes as unrefenced leading to
inconsistent state.

This PR initializes GC state when the container runtime loads, if a
current reference timestamp. Then every time there is a new connection,
if GC state is initialized, it updates the unreferened state based on
the current reference timestamp and if GC is not initialized, it
initializes it.
This ensures that any changes or loads to unreferenced nodes are
detected and that the GC sweep op will be processed after the GC state
has been initialized.

Note: This PR also converts the `TombstoneReadyObject_*` and
`SweepReadyObject_*` logs to generic instead of error. These are already
logged separately so they are mostly redundant. However, there are
certain scenarios where these are still useful such as in interactive
clients, when an object is not tombstone-ready / sweep-ready during
initialization but later becomes so.


AB#5760