build(merge-tree,sequence): Fix type generation for ESM and add exports field (#18825)
This PR updates sequence and merge-tree to generate proper types in ESM
builds and adds an exports field. I had to update the test files a bit
to clean up imports. I also added a ./dist/test export from merge-tree
since its internals are used by sequence tests. This is a demonstration
of the mechanism we will have to use any time we want to export
something for use in another package. No more reaching into internals!

This PR includes some configuration and settings changes that are not
strictly needed for this PR, but splitting them out doesn't seem worth
the time since they're all needed for other pending PRs like #18823 and