build(fluid-framework): Bundle map APIs via api-extractor (#18818)
Bundles exports from `@fluidframework/map` in `fluid-framework`. This
has a few benefits:
1. Makes the API surface of `fluid-framework` more readily apparent via
the generated API reports
    * See the changes made to `fluid-framework.api.md`
2. Ensures that we aren't missing required exports from the package
being bundled
* In this PR, api-extractor detected that we were missing `ICreateInfo`,
which is referenced by one of our existing exports.
3. Improves the generated API documentation by surfacing API members
directly on the page of the `fluid-framework` package, rather than
requiring that we point readers to the docs of another package.
* See [here](https://fluidframework.com/docs/apis/fluid-framework/) for
an example of the current hacky workaround we have in place on our
website to make the docs legible.

I'm tackling this work package-by-package, rather than bundling
everything here all at once. Other DDSs will be bundled in subsequent
PRs