Enhance sanitization logic w/ base common cases (#18639)
## Description

This change intends to prevent secrets getting leaked into the logs.
Modifications were made to sanitization formatter, so that known cases
were sensitive data is leaked gets formatted straight away, and that
sanitization logic is performed and consolidated inside the telemetry
library. This common cases are also performed before searching for
unknown cases when the sanitization is enabled.

In the future, the plan is to monitor for new detected cases, and add
them to the common cases when detected (which are always enabled and
performed in an efficient way).