Updated GC trailing ops test to check for deletion and other fixes (#18810)
The GC trailing ops test runs in two modes - tombstone enabled and
disabled. However, the tombstone / sweep part is not properly
implemented. The sweep timeout is 1ms (which is too small to be
meaningful) and the test doesn't wait for this timeout before or after
sending trailing ops which is contrary to the test title.

This PR changes the sweep timeout to 100ms, waits for this timeout
before or after sending trailing ops. It also validates that the data
store is not deleted when the container is opened after sweep timeout
and sweep phase runs.