build(devtools): Update package scopes to `@fluidframework` in preparation for upcoming 2.0 beta (#18820)
Renames the packages from the `@fluid-experimental` namespace to the
`@fluidframework` namespace in preparation for the upcoming 2.0 release
candidate.

Note that while only the `devtools` package will be advertised publicly,
it depends on `devtools-core`, so both really need to move.

Additionally, while we don't publish `devtools-view`, we do use it in
"production" as a part of our browser extension, so it also really
doesn't belong in `@fluid-experimental` anymore. This PR moves it to
`@fluid-internal`.