R11s: log response finish-to-close gap (#15139)
## Motivation

In AFR, we get a lot of reports about slow response times even when our
service telemetry reports more reasonable response times. This can range
from a couple hundred milliseconds to a minute or more. From a service
reliability perspective, we _need_ to get visibility there. This is an
**experimental** change to attempt to capture part of that story in
telemetry.

## Description

Node.js Responses have 2 events: `close` and `finish`. According to
Node.js documentation:

> Event: 'close'
[#](https://nodejs.org/docs/latest-v16.x/api/http.html#event-close_2)
> Added in: v0.6.7
> 
> Indicates that the response is completed, or its underlying connection
was terminated prematurely (before the response completion).

> Event: 'finish'
[#](https://nodejs.org/docs/latest-v16.x/api/http.html#event-finish_1)
> Added in: v0.3.6
> 
> Emitted when the response has been sent. More specifically, this event
is emitted when the last segment of the response headers and body have
been handed off to the operating system for transmission over the
network. It does not imply that the client has received anything yet.

[morgan](https://github.com/expressjs/morgan), the Express middleware we
use for logging HTTP Request duration, logs on the 'finish' event (or
'close' if premature termination) using the
[on-finished](https://www.npmjs.com/package/on-finished) package.

The goal here is to log the gap between request "finish" (hand-off to
OS) and "close" (response complete).

## Caveats

I don't actually know if this will work as intended. There are some gaps
in my knowledge about what is actually happening here. However, better
to have a little more telemetry than not, as we suspect Network I/O
could be causing bottlenecks in AFR.