tree: Basic schema editing support (#18815)
## Description

This PR introduces basic support for schema editing. It makes schema
edits a first-class concept, adding support for it in the SharedTree op
type and handling it in rebasing.

## Breaking Changes

This moves schema editing off of the stored schema and onto the view. It
also breaks the persisted format by changing the top-level op format.