build-tools: Add new tasks for build caching (#18886)
Makes several updates to fluid-build so that it caches results better.

- Updates the api-extractor task to include additional config files in
its cache key. This does mean that all api-extractor tasks are
invalidated if any of the config files change. However, it still doesn't
take into account the shared configs that the leaf configs inherit from.
So changes to the root config do not invalidate the cache. Adding that
is a larger change.
- Adds an option to the LeafWithFileStatDoneFileTask that enables the
DoneFile it creates to use file contents hashes instead of file stats.
This is useful for tasks whose metadata may be legitimately updated
during the build. This is not currently used, but it will likely be
needed for some upcoming tasks.
- Adds a new flub command, `rename-types`, that will replace renamer
because renamer complains loudly when there are no source files, and it
needs to run without issue in that scenario.