Gitrest: refactor WholeSummaryManager (#18221)
## Description

Gitrest's `GitWholeSummaryManager` file is almost 1000 lines long, and
is incredibly difficult to parse through. We have been seeing memory and
performance issues in Gitrest, so this is the start of attempting to
break down this massive code chunk in the name of maintainability:

- Read
- Write
- Core utils (shared across low-io and normal)
- Low-io utils (things only done for low-io mode)
- Conversions (all the type/shape conversions that happen within a lot
of the logic)

I also added unit testing for:
- Storage Access counts
- Storage sizes
- Cross compatibility for low-io and normal modes

## Reviewer Guidance

- There are no major logic changes happening in this PR, just breaking
the `gitWholeSummaryManager` file apart into `definitions`, `constants`,
`conversions`, `readWholeSummary`, and `writeWholeSummary`.
- The only real change happening is that I am passing `blobCache` around
to a few more places that previously did not use it. This very slightly
reduces the calls to storage.
- The next step will be utilizing these more logical and declarative
codepaths to add improved logging around the performance of individual
pieces of the Gitrest summary write process.