GC: Rename "sweep timeout" to "tombstone timeout" throughout (#18852)
PR #18559 split the old "sweep timeout" into two segments: "tombstone
timeout" followed by "sweep grace period". The new "tombstone timeout"
is equivalent to the old "sweep timeout", so to keep that diff smaller I
avoided renaming it in most places in that PR.

This PR does the rename everywhere. Most of the change is mechanical,
but in `gcConfigs.ts` there is some back-compat code since
`IGCMetadata.sweepTimeoutMs` is persisted. So as we rename it to
`tombstoneTimeoutMs`, we also need to continue to check for
`sweepTimeoutMs` when reading.

## Breaking Changes

It's technically a change to the snapshot format, but done in an
incremental way. Snapshot tests didn't break, because (apparently)
sessionExpiry is disabled and so sweepTimeoutMs is not present in the
reference files.