build: Build ESM type rollups using api-extractor and update tsc-multi patch (#18903)
Prior to this change, the ESM types for a package were copied from the
CJS output. This is this wrong in many cases, and so this PR updates the
ESM api-extractor tasks to call it directly. I added a new shared config
in build-common, but due to how api-extractor determines the paths, each
project still needs to have its own local config.

This change also updates the tsc-multi patch to rename declaration files
and correct source map references. Because tsc-multi now renames the
files, we can remove the rename task and dependency.