build: Don't distribute test files generated during build (#18914)
I noticed our recent merge-tree release was ~9MB with an unpacked size
of over 47MB. The main culprit is the inclusion of test files, which is
explained in a comment in the npmignore file. However, there is a tar
file generated during testing that we don't need to distribute, so I
added an entry to ignore that file to all packages that generate one.