Fix several tree2 transaction issues (#18689)
## Description

Fixes several issues related to transactions in tree2. This change is
motivated by work to support transactions in optional-field, or at least
anticipate places where they might necessitate changes to the persisted
format. To that end, I started writing some axiomatic tests before
realizing I wasn't totally sure how the production codepath worked with
respect to ID allocation and several other concerns. Turns out it had
some bugs at the branch level:

- AB#6484: Transactions starting on a point of divergence inadvertently
included commits that they were rebased over. This resulted in various
sorts of errors. This PR introduces a fix in `branch.ts` and adds some
unit tests for the issue.
- AB#6485: The general rebase scheme does not preserve transactionality.
In addition to the work item, I left some comments to this effect on the
transaction-relevant private state in `Branch`. This PR does not attempt
a fix for this issue.

The PR also adjusts the implementation for optional field's `invert`,
which previously didn't support changes which might be made using
transactions. Some tests for relevant scenarios are added to
`editing.spec.ts` for that. This fixes some issues in previously skipped
fuzz seeds.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>