GC: Added whether tombstone is enabled to GarbageCollection_end event (#18927)
Added a `tombstone` property to GarbageCollection_end event to indicate
whether throwing on tombstone is enabled for a GC run. This will help to
filter out sessions / GC runs where tombstone did not run. The event
already had `sweep` property to indicate whether sweep ran.

Also, update the `PerformanceEvent` class to use
`ITelemetryGenericEventExt` and `ITelemetryPropertiesExt`. This allowed
the property `details` in the GarbageCollection_end event to be an
object.

AB#6397