Added test that validates GC state is updated before trailing ops are processed (#18865)
#18758 update the logic to initialize GC state. This used to happen when
container connects in "write" mode but now it happens during GC
(container runtime) initialization. So, any trailing ops (or any ops for
that matter) before "write" mode that use unreferenced objects will
result in logging an error log.

Basically, before #18758, GC would miss logging unreferenced errors
before container connected in "write" mode and now it won't. This PR
adds a unit test to validate that. In the test, GC initializes with an
object that is inactive, tombstone or sweep ready. This objects is then
loaded, receives an op and is revived and the test validates that the
appropriate events are logged.
Also, added clientId to `GCInitializationOrUpdateFailed` event and a
comment that was missed in #18758.

AB#6606