GC: Fix GC deleted stats not correctly updated for data stores (#18890)
## Bug
#18506 added deleted stats to the GarbageCollection_end logs. However,
the calculation for getting deleted data stores count is incorrect. It
calls `this.runtime.getNodeType(nodeId)` to get the node's type which
checks if the data store with the given nodeId exists. Since the data
store has been deleted, it ends up returning the type as other.

## Fix
In case of deleted nodes, the garbage collector gets the node type
itself by looking the path. This should be fine because it is only used
for stats and for the current node types (data store, DDS and blob),
this will work fine.