Finalize IDs immediately when attaching a detached container (#18945)
## Description
Prior to this if a client generated IDs using the `IdCompressor` while
detached it could end up serializing un-finalized IDs in the attach
summary. Since we're the only client, we don't need to wait on consensus
and can immediately claim a cluster and finalize the range.

There's not test in this PR, but `SharedTree` has a PR following this up
which has test failures without this change. Writing a unit test here to
test this is a part of
https://dev.azure.com/fluidframework/internal/_workitems/edit/6675 as
follow-up to this work

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>