Add ghost compressor API to support migration scenario (#18960)
## Description
Using the compressor in the new SharedTree was causing the migration
tests to fail. We added an API to the compressor to support the
migration scenarios. As part of this PR we're enabling the compressor in
migration tests that will need it when tree merges its use of the
compressor.